refactor(api): extract shared user service
