+++ apps/api/src/user-service/user-service.module.ts
@@ -1,0 +1,10 @@
+import { Module } from "@nestjs/common";
+import { PrismaModule } from "../prisma/prisma.module.js";
+import { UserService } from "./user.service.js";
+
+@Module({
+  imports: [PrismaModule],
+  providers: [UserService],
+  exports: [UserService],
+})
+export class UserServiceModule {}
