+++ apps/api/src/users/users.module.ts
@@ -1,8 +1,10 @@
 import { Module } from "@nestjs/common";
+import { UserServiceModule } from "../user-service/user-service.module.js";
 import { UsersController } from "./users.controller.js";
 import { UsersService } from "./users.service.js";
 
 @Module({
+  imports: [UserServiceModule],
   controllers: [UsersController],
   providers: [UsersService],
   exports: [UsersService],
